requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.0
blis==0.7.11
boto==2.49.0
boto3==1.35.4
botocore==1.35.4
catalogue==2.0.10
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.15.1
colorama==0.4.6
confection==0.1.5
cymem==2.0.8
dataclasses-json==0.6.7
distro==1.9.0
Django==4.2.4
django-cors-headers==4.4.0
django-crontab==0.7.1
django-environ==0.11.2
django-rest-passwordreset==1.4.1
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-api-tracking==1.8.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl#sha256=6215d71a3212690e9aec49408a27e3fe6ad7cd6c715476e93d70dc784041e93e
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.4.0
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.34
langchain-openai==0.1.22
langchain-text-splitters==0.2.2
langcodes==3.4.0
langsmith==0.1.104
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.24.0
openai==1.42.0
orjson==3.10.7
packaging==24.1
pandas==2.0.1
pathlib_abc==0.1.1
pathy==0.11.0
pillow==10.4.0
preshed==3.0.9
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-environ==0.4.54
pytz==2024.1
PyYAML==6.0.2
razorpay==1.4.2
regex==2024.7.24
requests==2.32.3
rich==13.7.1
s3transfer==0.10.2
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.32
sqlparse==0.5.1
srsly==2.4.8
tenacity==8.5.0
thinc==8.2.1
tiktoken==0.7.0
tqdm==4.66.5
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.16
vine==5.1.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.3.2
wrapt==1.16.0
yarl==1.9.4
